requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
phonenumbers==8.13.16
